CM-2101: Renamed DataModelUtils into RestApiUtils for better clarity. Refactored to move creation of request parameters into RestApiUtils.
