+++ comet-java-client/src/main/java/ml/comet/experiment/impl/utils/ModelUtils.java
@@ -1,17 +1,6 @@
 package ml.comet.experiment.impl.utils;
 
-import lombok.NonNull;
-import ml.comet.experiment.impl.constants.QueryParamName;
-import ml.comet.experiment.registrymodel.DownloadModelOptions;
 import org.apache.commons.lang3.StringUtils;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import static ml.comet.experiment.impl.constants.QueryParamName.MODEL_NAME;
-import static ml.comet.experiment.impl.constants.QueryParamName.STAGE;
-import static ml.comet.experiment.impl.constants.QueryParamName.VERSION;
-import static ml.comet.experiment.impl.constants.QueryParamName.WORKSPACE_NAME;
 
 /**
  * The collection of utilities to work with registry models.
@@ -59,25 +48,4 @@ public class ModelUtils {
         return modelName;
     }
 
-    /**
-     * Creates query parameters to be used to download model from the Comet registry.
-     *
-     * @param workspace    the name of the model's workspace.
-     * @param registryName the model's name in the registry.
-     * @param options      the additional download options.
-     * @return the map with query parameters.
-     */
-    public static Map<QueryParamName, String> downloadModelParams(
-            @NonNull String workspace, @NonNull String registryName, @NonNull DownloadModelOptions options) {
-        Map<QueryParamName, String> queryParams = new HashMap<>();
-        queryParams.put(WORKSPACE_NAME, workspace);
-        queryParams.put(MODEL_NAME, registryName);
-        if (StringUtils.isNotBlank(options.getVersion())) {
-            queryParams.put(VERSION, options.getVersion());
-        }
-        if (StringUtils.isNotBlank(options.getStage())) {
-            queryParams.put(STAGE, options.getStage());
-        }
-        return queryParams;
-    }
 }
